refactor user/hooks
