+++ packages/core/src/features/user/utils.ts
@@ -1,5 +1,8 @@
 import { coreStore } from '../../store';
 import { userAuthApi } from './userSliceRTK';
+import { fetchJSONDataFromURL, HttpMethod } from '../../utils';
+import { GEN3_WTS_API } from '../../constants';
+import { ExternalProvider } from './types';
 
 let cachedToken: string | null = null;
 
@@ -72,4 +75,38 @@ export const makeAuthenticatedRequest = async <T>(
   }
 
   return response.json() as Promise<T>;
+};
+
+interface ExternalProviders {
+  providers: ExternalProvider[];
+}
+
+interface ExternalProvidersResponse extends ExternalProviders {
+  error?: Error;
+}
+
+export const queryWTSFederatedLoginStatus = async (
+  signal?: AbortSignal,
+): Promise<ExternalProvidersResponse> => {
+  try {
+    const results = await fetchJSONDataFromURL<ExternalProviders>(
+      `${GEN3_WTS_API}/external_oidc/`,
+      false,
+      HttpMethod.GET,
+      undefined,
+      signal,
+    );
+
+    // Handle both null response and response without providers
+    if (!results || !results.providers) {
+      return { providers: [] };
+    }
+
+    return results;
+  } catch (error: unknown) {
+    return {
+      providers: [],
+      error: error instanceof Error ? error : new Error('Unknown error'),
+    };
+  }
 };
